1.10.0 - enabled proprietary extensions.
A mechanism to extend the command sets is now added.
The pattern is to make a module containing a set of command classes that derive from BasicCommand, plus a KnownCommands() function.

Then you change your run script to include the new sets of commands, noting that
the commands are parsed in order, so if a command of the same name is implemented later, it is used in preference.

The code pattern to enable the additional commands to be used is:
extended_command_sets = []
extended_command_sets.append('distpy_scr.calc.slb_command_set')
distpy.controllers.parallel_strainrate_processing.main(configStrainrate2Summary, extended_list=extended_command_sets)